+++ packages/patrol/CHANGELOG.md
@@ -5,6 +5,8 @@
   To do so, pass `test-server-port` (default 8081) and `app-server-port`
   (default 8082) to Patrol CLI commands (e.g `patrol test`)
 
+- Fix `patrolSetUp()` and `patrolTearDown()` not working in `patrol develop` (#2023).
+
 This version requires version 2.6.0 of `patrol_cli` package.
 
 ## 3.3.0